Moved getTagById() fetch call inside openDialog() to resolve lag issue
