+++ src/services/index.js
@@ -29,3 +29,6 @@ export { getAllReactions } from "./ReactionService.js";
 
 // PostReactions
 export { getAllPostReactions, getPostReactionsByPostId, updateOrCreatePostReaction } from "./PostReactionService.js";
+
+//Tags
+export { getAllTags, createTag, getTagById, editTag } from "./TagService.js";
